+++ R/calc.R
@@ -56,7 +56,7 @@ Bacon.Age.d <- function(d, set=get('info'), its=set$output, BCAD=set$BCAD, na.rm
          ages <- NA # Feb 2021
 
     if(BCAD)
-      ages <- 1950 - ages
+      ages <- rice::calBPtoBCAD(ages)
     return(c(ages))
 }
 
@@ -85,8 +85,8 @@ Bacon.Age.d <- function(d, set=get('info'), its=set$output, BCAD=set$BCAD, na.rm
 #' @export
 Bacon.d.Age <- function(age, set=get("info"), BCAD=set$BCAD, its=set$output, na.rm=FALSE ){
 
-  if(BCAD) 
-    age <- 1950 - age # if the customer decides to enter the age in BCAD
+  if(BCAD && !set$BCAD) 
+    age <- rice::BCADtocalBP(age)
   if(length(age) > 1) 
     stop("Bacon.d.Age can handle one age at a time only", call. = FALSE)
   if(length(its) == 0 )